src/common_amirainvest_com/pyproject.toml
[build-system]
requires = ["setuptools", "poetry-core @ git+https://github.com/python-poetry/poetry-core.git@master"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "common_amirainvest_com"
version = "0.1.0"
description = ""
authors = ["Peter Hoburg <peterHoburg@users.noreply.github.com>"]
packages = [
    {include = "common_amirainvest_com", from="lib"}
]

[[tool.poetry.source]]
name = "amirapypi"
url = "https://amira-pypi-903791206266.d.codeartifact.us-east-1.amazonaws.com/pypi/amira_pypi/simple/"
default = true

[tool.poetry.dependencies]
asyncpg = "^0.25.0"
boto3 = "^1.20.24"
boto3-stubs = {extras = ["sqs"], version = "^1.20.24"}
fastapi = "^0.70.0"
psycopg2-binary = "^2.9.2"
pydantic = "^1.8.2"
pydantic-sqlalchemy = "^0.0.9"
python = "^3.9"
python-dotenv = "^0.19.2"
redis = "^4.0.2"
SQLAlchemy = "^1.4.27"

[tool.poetry.dev-dependencies]

